# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Mobile Restaurant Management API
  version: 1.1.0
  description: |
    REST API cho hệ thống đặt món — JWT, Socket.IO, tìm nhà hàng theo GPS, VNPay, Cloudinary.

    **Base path:** `/api` — Swagger UI: `/api-docs`

    **Roles:** `customer`, `restaurant_owner`, `delivery_staff`, `admin`
    - Đăng ký API chỉ cho `customer`, `restaurant_owner`, `delivery_staff`
    - Tài khoản `admin` cần tạo trong MongoDB (hoặc đổi role user có sẵn)

    **Luồng test gợi ý:**
    1. Owner đăng ký → tạo nhà hàng (status `pending`)
    2. Admin duyệt nhà hàng (`PATCH .../approve`)
    3. Owner tạo menu → Customer đặt đơn → thanh toán COD/VNPay
    4. Owner cập nhật trạng thái đơn theo flow: pending → confirmed → preparing → ready → delivering → completed

    **Socket.IO** (kết nối kèm `auth.token` = access JWT):
    `order_created`, `order_confirmed`, `order_preparing`, `order_ready`, `order_delivering`, `order_completed`, `order_cancelled`, `new_review`, `notification`
servers:
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/menu/{restaurantId}/foods:
    get:
      tags: [Menu]
      summary: Danh sách món
      security: []
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/menu/foods:
    post:
      tags: [Menu]
      summary: Tạo món (owner, multipart)
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [restaurantId, categoryId, name, price]
              properties:
                restaurantId:
                  type: string
                categoryId:
                  type: string
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: OK
  /api/menu/foods/{id}:
    put:
      tags: [Menu]
      summary: Cập nhật món (owner)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                isAvailable:
                  type: boolean
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
...